Create Kind Message
A program that asks the user for a kind message.  The program will then ask the user to verify the message and ask for a file to output the message.  The message should then be output to the end of the “kind message” file so that all of the messages are stored in one text document.